src/vector_store.py
import faiss
import numpy as np

def create_vector_store(embeddings):

    dimension = len(embeddings[0])

    index = faiss.IndexFlatL2(dimension)

    index.add(np.array(embeddings))

    return index