requirements.txt
langchain
ollama
faiss-cpu
pypdf
tiktoken
streamlit
python-dotenv
sentence-transformers
pandas